// src/swagger/model/project.ts
/**
 * Here to Help
 * This is a restful web service used to log hours for non-profits to submit for money grants.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Project {
    projectId: string;
    name: string;
    description?: string;
    startDate: Date;
    endDate: Date;
}